+++ Runtime/Levels/LevelData.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+namespace Pixygon.Micro
+{
+    public class LevelData : ScriptableObject {
+        public GameObject _levelPrefab;
+        public ParallaxLayerData[] _parallaxLayerDatas;
+    }
+}